Improve account block validation
Significant performance issues start appearing when the node validates an account block that uses an old momentum acknowledgement height.

The reason for the slow processing time of such an account block is that in order to verify the account block, the node must rollback the chain database's state to the block's momentum acknowledgement height. The rollback operation is done in the `Get` function in `versioned_db.go`.

In order to prevent the need for an expensive rollback operation, this commit introduces an additional chain cache database that stores the necessary state information in a way that can be accessed quickly.

To validate a *user* account block, the following stateful information is needed:

1. The amount of plasma fused to the sender's address at momentum acknowledgement height.
2. The amount of total chain plasma committed by the sender at momentum acknowledgement height.
3. What sporks are active at momentum acknowledgement height.

These three data items are stored in the cache using a momentum height composite key, allowing the key-value pairs to be iterated over in order to get the value at a specific height. Using the cache removes the need to rollback the momentum store to the momentum acknowledgment height, when validating a user account block.

To validate an *embedded* account block, various stateful data items are needed, since processing an embedded account block executes contract functionality. This means that the validation of an embedded account block still rolls back the momentum store to the momentum acknowledgement height. An embedded account block uses the send block's confirmation height as its acknowledgement height, which means that the amount of momentums needed to be rolled back will always be small, minimizing potential issues related to performance.

**Note:** An account block is not allowed to reference an older momentum acknowledgement height than what the previous account block has used. Because of this, the old states related to plasma are pruned. This significantly reduces the cache's size.