Improve account pool performance
The account pool starts experiencing performance issues when it holds a large amount of transactions. Specifically the account pool's (account_pool.go) functions getUncommittedAccountBlocksByAddress and rebuild become unacceptably slow once there are more than a few hundred pending account blocks in the pool.

The reason for the slowness is the multiple consecutive read operations done to LevelDB's in-memory database (memdb) in the getUncommittedAccountBlocksByAddress and rebuild functions. The memdb's performance is not adequate for this use case.

To fix this issue, this commit introduces an accountManager which is responsible for maintaining a key-value map of an account's pending blocks. This way the pending blocks do not have to be fetched from the memdb, significantly improving performance.

Additionally, when the account pool is rebuilt, the account manager databases are now reused, instead of being rebuilt from scratch. This further improves performance.

With these changes, the account pool is able to maintain a pool of hundreds of thousands of pending account blocks without performance issues.