Improve account pool performance
The account pool starts experiencing performance issues when it holds a large amount of account blocks from account chains that have long unconfirmed states. Specifically the account pool's (`account_pool.go`) functions `getUncommittedAccountBlocksByAddress` and `rebuild` become unacceptably slow if there are more than a few account chains with hundreds of pending account blocks on each chain.

The reason for the slowness is the multiple consecutive read operations done to the account chain's unconfirmed (frontier) database in the `getUncommittedAccountBlocksByAddress` and `rebuild` functions. The account chain's unconfirmed database's performance is not adequate for this use case.

To fix this issue, this commit introduces an `accountManager` which is responsible for maintaining a key-value map of an account's pending blocks. This way the pending blocks do not have to be fetched from the unconfirmed account database, significantly improving performance.